Refactor return types on async calls from interfaces to concrete types, to reduce errors. This will be changed after deciding on specific weather interfaces to provide.
