+++ WundergroundNetLib/Data/WundergroundData.cs
@@ -6,16 +6,16 @@ using System.Threading.Tasks;
 
 using WundergroundNetLib.Interfaces;
 
-namespace WundergroundNetLib
+namespace WundergroundNetLib.Data
 {
-    public class WundergroundData : IWundergroundData
+    public class WundergroundData
     {
-        public IObservationLocation observationLocation { get; set; }
-        public ICurrentConditions currentConditions { get; set; }
-        public List<IForecast> fourDayForecast { get; set; }
+        public ObservationLocation observationLocation { get; set; }
+        public CurrentConditions currentConditions { get; set; }
+        public List<Forecast> fourDayForecast { get; set; }
     }
 
-    public class ObservationLocation : IObservationLocation
+    public class ObservationLocation
     {
         public string City { get; set; } // city : "Christchurch" / display_location
         public string Country { get; set; } // country_iso3166 : "NZ" / observation_location
@@ -26,7 +26,7 @@ namespace WundergroundNetLib
         public int WmoNumber { get; set; } // wmo : "93780" / display_location
     }
 
-    public class CurrentConditions : ICurrentConditions
+    public class CurrentConditions
     {
         public DateTime ObservationTime { get; set; } // observation_time_rfc822 : "Fri, 30 Oct 2015 07:56:47 +1300"
         public string CurrentDescription { get; set; } // weather : "Rain"
