+++ WundergroundNetLib/Model/WundergroundDataProvider.cs
@@ -1,8 +1,8 @@
 ﻿using System;
 using System.Threading.Tasks;
 using Newtonsoft.Json;
-
 using WundergroundNetLib.Interfaces;
+using WundergroundNetLib.Data;
 
 namespace WundergroundNetLib
 {
@@ -34,7 +34,7 @@ namespace WundergroundNetLib
         /// <param name="location"></param>
         /// <param name="dataFeatures"></param>
         /// <returns></returns>
-        public async Task<IWundergroundData> GetWundergroundWeatherDataAsync(string latitude, string longitude)
+        public async Task<WundergroundData> GetWundergroundWeatherDataAsync(string latitude, string longitude)
         {
             UriProvider uriProvider = new UriProvider();
             Uri pwsUri = uriProvider.CreateCombinedDataUriFromCoordinates(latitude, longitude);
@@ -48,7 +48,7 @@ namespace WundergroundNetLib
         /// <param name="latitude"></param>
         /// <param name="longitude"></param>
         /// <returns></returns>
-        public async Task<IWundergroundData> GetWundergroundWeatherDataAsync(double latitude, double longitude)
+        public async Task<WundergroundData> GetWundergroundWeatherDataAsync(double latitude, double longitude)
         {
             UriProvider uriProvider = new UriProvider();
             Uri pwsUri = uriProvider.CreateCombinedDataUriFromCoordinates(latitude, longitude);
@@ -61,7 +61,7 @@ namespace WundergroundNetLib
         /// </summary>
         /// <param name="stationID"></param>
         /// <returns></returns>
-        public async Task<IWundergroundData> GetWundergroundWeatherDataAsync(string stationID)
+        public async Task<WundergroundData> GetWundergroundWeatherDataAsync(string stationID)
         {
             UriProvider uriProvider = new UriProvider();
             Uri pwsUri = uriProvider.CreateCombinedDataUriFromPwsStationID(stationID);
@@ -73,7 +73,7 @@ namespace WundergroundNetLib
         /// </summary>
         /// <param name="pwsUri"></param>
         /// <returns></returns>
-        private async Task<IWundergroundData> CombinedWeatherDataAsync(Uri pwsUri)
+        private async Task<WundergroundData> CombinedWeatherDataAsync(Uri pwsUri)
         {
             // Download Json data
             JsonProvider jsonProvider = new JsonProvider();
